src/requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.81.0
h11==0.13.0
httptools==0.4.0
idna==3.3
pydantic==1.9.2
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.16.1
websockets==10.3
google-cloud-gke-hub==1.5.0
google-cloud-logging==3.2.1
google-cloud-compute==1.5.1